Changed the amount type to have decimal numbes. Before it was only integers and expense amounts are not expected to be integers.

+++ src/backend/prisma/migrations/20250919151919_expense_backend/migration.sql
@@ -1,7 +1,7 @@
 -- CreateTable
 CREATE TABLE `Expense` (
     `id` BIGINT NOT NULL AUTO_INCREMENT,
-    `amount` INTEGER NOT NULL,
+    `amount` DECIMAL(10, 2) NOT NULL,
     `date` DATETIME(3) NOT NULL,
     `category` ENUM('Food', 'Groceries', 'Mobile_Bill', 'Travel', 'Shopping', 'Games', 'Subscription', 'EMI') NOT NULL,
     `description` VARCHAR(191) NULL,
